Updated ioptron_mount to support 4 children
This matches with 4 of the planes where we might
want to place weatherproofing. Could add another
(the "non-moving" side of the DEC axis).

Documented and enforced a coordinate system for
each spot where children can be placed.

Added a bearing gap to the DEC axis.

Added example objects that show the placement
of each child.