Add a fifth child to module ioptron_mount
Allows attachment of weatherproofing to the DEC      
worm cover/DEC motor, rather than to the DEC
head on the otherside of the bearing gap, which
moves as the DEC shaft rotates. Not yet sure
if the coordinate system of thhat new child is
appropriate.

Supported the above by adding a fourth child
to module ra_to_dec

Cleanup some formatting and echo messages,
added some more comments.