Add Polar Scope port to top of DEC axis
Add knobs for securing a dovetail to the saddle plate.
Remove or comment out a bunch of echo calls.
Remove some extra axis decorations in the sample for ra_and_dec.
Add shoulder for RA to DEC2 attachment.

Changed approach used to generate dec_head_bearing_cover so that
there are fewer artifacts in the preview produced by OpenSCAD.