WIP on a new, simpler helmet.
This is a can, open bottom, lid on top. Lid printed separately
from the can; can printed in 4 parts to avoid 3D printing problems
with overhangs.

wp5/shaft_port.scad is used for both DEC head and CW shaft ports;
it has a 45 degree slope to the part that is an overhang, allowing
it to be printed reasonably easily.