Added bosses to support screwing the two halves of the helmet together.
Added wp5/printers/* containing .scad files with models to
be turned into STL files, along with a script for processing
all of the .scad files in the directory.

Adapted dec_head_bearing_cover.scad from weatherproofing_2,
inverting it for printability, and adding slots for AWG 12
wires, for an experiment in making a DIY slip ring.